requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2
requests
Flask-Uploads
Flask-Migrate
python-dotenv
